Create initial project structure
Lay out the foundation for which this project will build upon.

Add POM file to the root of the project to leverage the Maven build
tool. This will allow for easy consistent builds as the project grows.

Add .gitignore file to avoid anyone commiting files that do not belong
to the project, like compiled JARs files.

Add a simple App.java file with a simple main method where we will start
to add code

Add a simple AppTest.java file where we will add tests later on.

Signed-off-by: Anthony M. Bonafide <AnthonyMBonafide@gmail.com>